requirements.txt
streamlit==1.35.0
feedparser==6.0.10
google-generativeai==0.3.0
SpeechRecognition==3.10.0
pydub==0.25.1
edge-tts==6.1.1
gTTS==2.4.0
python-dotenv==1.0.0
pyaudio==0.2.13
